#335: Moved all objets-display-related components to a new folder
 Let's gather related components.
