+++ src/app/ffbe/objets-display-components/objet-alterations-etat-display/objet-alterations-etat-display.component.html
