+++ src/app/ffbe/objets-display-components/objet-alterations-etat-duo-display/objet-alterations-etat-duo-display.component.html
