+++ src/app/ffbe/objets-display-components/objet-carac-display/objet-carac-display.component.html
