+++ src/app/ffbe/objets-display-components/objet-display/objet-display.component.css
