+++ src/app/ffbe/objets-display-components/objet-display/objet-display.component.html
@@ -1,8 +1,8 @@
 <mat-card>
   <mat-card-title (click)="switchDisplayed()">
-    <img *ngIf="present && !different" src="assets/ok.png" class="ffchstatus" alt="OK">
-    <img *ngIf="!present" src="assets/nok.png" class="ffchstatus" alt="Not OK">
-    <img *ngIf="present && different" src="assets/grey/warn.png" class="ffchstatus" alt="Warn">
+    <img *ngIf="present && !different" src="../../../../assets/ok.png" class="ffchstatus" alt="OK">
+    <img *ngIf="!present" src="../../../../assets/nok.png" class="ffchstatus" alt="Not OK">
+    <img *ngIf="present && different" src="../../../../assets/grey/warn.png" class="ffchstatus" alt="Warn">
     {{objet.gumi_id}} {{objet.nom}}
   </mat-card-title>
 
@@ -57,10 +57,10 @@
             <img *ngIf="objet.isImagePresentInFfchDb()" src="{{objet.icone}}"
                  alt="{{objet.nom}}" title="{{objet.nom}} ({{objet.gumi_id}})"/>
             <img
-              *ngIf="!objet.isImagePresentInFfchDb()"
-              src="assets/grey/warn.png" alt="{{objet.nom}} ({{objet.gumi_id}})"
-              title="{{objet.nom}} ({{objet.gumi_id}})"/>
-            <img *ngIf="!objet.isPresentInFfchDb()" src="assets/red/warn.png"
+                    *ngIf="!objet.isImagePresentInFfchDb()"
+                    src="../../../../assets/grey/warn.png" alt="{{objet.nom}} ({{objet.gumi_id}})"
+                    title="{{objet.nom}} ({{objet.gumi_id}})"/>
+            <img *ngIf="!objet.isPresentInFfchDb()" src="../../../../assets/red/warn.png"
                  alt="{{objet.gumi_id}}" title="{{objet.gumi_id}}"/>
           </span>
         </td>
