+++ src/app/ffbe/objets-display-components/objet-elements-display/objet-elements-display.component.html
