+++ src/app/ffbe/objets-display-components/objet-elements-triplet-display/objet-elements-triplet-display.component.css
