+++ src/app/ffbe/objets-display-components/objets-display/objets-display.component.css
