+++ src/app/ffbe/objets-display-components/objets-display/objets-display.component.ts
@@ -1,10 +1,10 @@
 import {Component, Input, OnChanges, OnDestroy} from '@angular/core';
-import {Objet} from '../model/objet/objet.model';
+import {Objet} from '../../model/objet/objet.model';
 import {forkJoin, Observable, of, Subscription} from 'rxjs';
-import {ObjetsComparingContainer} from '../model/objet/objets-comparing-container.model';
-import {FfchClientService} from '../services/ffch-client.service';
+import {ObjetsComparingContainer} from '../../model/objet/objets-comparing-container.model';
+import {FfchClientService} from '../../services/ffch-client.service';
 import {catchError} from 'rxjs/operators';
-import {FfbeUtils} from '../utils/ffbe-utils';
+import {FfbeUtils} from '../../utils/ffbe-utils';
 
 @Component({
   selector: 'app-objets-display',
